Move RsaAuthenticator to common package
