+++ game-core/src/test/java/games/strategy/engine/lobby/server/db/UserControllerIntegrationTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test;
 import org.mindrot.jbcrypt.BCrypt;
 import org.triplea.test.common.Integration;
 
-import games.strategy.engine.lobby.server.login.RsaAuthenticator;
+import games.strategy.engine.lobby.common.login.RsaAuthenticator;
 import games.strategy.engine.lobby.server.userDB.DBUser;
 import games.strategy.util.Md5Crypt;
 import games.strategy.util.Util;
